requirements.txt
# Core ML/DL packages
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Database packages
sqlalchemy>=1.4.0
pymysql>=1.0.2
psycopg2-binary>=2.9.0
sshtunnel>=0.4.0

# Data processing
pandas>=1.3.0
pillow>=8.3.0
opencv-python>=4.5.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and utilities
python-dotenv>=0.19.0
requests>=2.26.0

# Optional packages for enhanced functionality
wandb>=0.12.0  # For experiment tracking
tensorboard>=2.7.0  # For visualization
grad-cam>=1.3.0  # For interpretability

# Development and testing
pytest>=6.2.0
jupyter>=1.0.0
